Migrate SCM polymorphisms into classes
The way the code was laid out can work but it is not "pythonic"; in
Python, modules are not "first class citizen" and they can not be
really manipulated like variables (some edge cases can bite you). It
is safer to declare classes with static methods which will simply act
as namespaces.