Try fix some parts of the generated html
The generated html does not comply with HTML specification and relies
on browser's good will to self heal ill-formed documents.

This change still leaves some parts of the HTML file with validation
errors, because the way it is generated, by regexp replacement,
instead of full parsing of the source, is quite coarse and brittle.

Until a real parser is implemented, this is a least error prone
replacement system I could come up with.